+++ modules/ui.lua
@@ -152,3 +152,67 @@ function GigaSettingsInterface:UpgradeRaidFrames()
     end)
 	end
 end
+
+local function GetPlayerMapCoords()
+  local uiMapId = C_Map.GetBestMapForUnit("player")
+  if (uiMapId == nil) then
+    return ""
+  end
+  local xy = C_Map.GetPlayerMapPosition(uiMapId, "player")
+  if (xy == nil) then
+    return ""
+  else
+    local x, y = xy:GetXY()
+    return format("%.2f, %.2f",x*100, y*100)
+  end
+end
+local playerMinimapCoordsTicker = nil
+function GigaSettingsInterface:PlayerMinimapCoords()
+  if not GigaSettingsDB.playerMinimapCoords or IsInInstance() then
+    if playerMinimapCoordsTicker then playerMinimapCoordsTicker:Cancel() Minimap.GigaPlayerCoords:SetText("") end
+    playerMinimapCoordsTicker = nil  
+    return
+  end
+
+  if not Minimap.GigaPlayerCoords then
+    Minimap.GigaPlayerCoords = Minimap:CreateFontString(nil, "OVERLAY", "GameFontHighlightSmall")
+    Minimap.GigaPlayerCoords:SetPoint("TOP", Minimap, "TOP", 0, -12)
+  end
+  playerMinimapCoordsTicker = C_Timer.NewTicker(0.3, function()
+      Minimap.GigaPlayerCoords:SetText(GetPlayerMapCoords())
+    end)
+end
+
+local cursorRingTicker = nil
+function GigaSettingsInterface:CursorRing()
+  if not GigaSettingsDB.cursorRing then
+    if cursorRingTicker then cursorRingTicker:Cancel() UIParent.GigaCursorRing:Hide() end
+    cursorRingTicker = nil  
+    return
+  end
+
+  if not UIParent.GigaCursorRing then
+    UIParent.GigaCursorRing = UIParent:CreateTexture(nil, "ARTWORK")
+    UIParent.GigaCursorRing:EnableMouse(false)
+    UIParent.GigaCursorRing:SetPoint("CENTER")
+    UIParent.GigaCursorRing:SetAtlas(GigaSettingsDB.cursorRingTexture)
+    UIParent.GigaCursorRing:SetSize(100, 100)
+  end
+  UIParent.GigaCursorRing:Show()
+  local ringAtlas = UIParent.GigaCursorRing:GetAtlas()
+  if ringAtlas and ringAtlas ~= GigaSettingsDB.cursorRingTexture then
+    UIParent.GigaCursorRing:SetAtlas(GigaSettingsDB.cursorRingTexture)
+    return
+  end
+
+  cursorRingTicker = C_Timer.NewTicker(0.01, function()
+      local x, y = GetCursorPosition()
+      local scale = UIParent:GetEffectiveScale()
+      UIParent.GigaCursorRing:SetPoint("CENTER", UIParent, "BOTTOMLEFT", x / scale, y / scale)
+    end)
+  -- UIParent.GigaCursorRing:SetScript("OnUpdate", function(self)
+  --   local x, y = GetCursorPosition()
+  --   local scale = UIParent:GetEffectiveScale()
+  --   self:SetPoint("CENTER", UIParent, "BOTTOMLEFT", x / scale, y / scale)
+  -- end)
+end